refactor: extract skip function
